add support for socket timeouts
this enables a configuration option for `timeout` that defaults to 5 seconds. this is used to give a reasonable timeout on the socket connection. without it, the timeout appears to be more than 60 seconds. for a plugin that is monitoring a large number of ephemeral hosts, the default timeout is too long.